+++ tests/test_konno_ohmachi.py
@@ -7,8 +7,6 @@ import pykooh
 
 from . import DATA_PATH
 
-use_cython = [False, True]
-
 # Load test data generated from the example.ipynb notebook.
 # The reference is computed by Obspy
 trace = obspy.read(DATA_PATH / "example_ts.mseed").traces[0]
@@ -19,18 +17,13 @@ bw = 40
 fa_sm_obspy = konno_ohmachi_smoothing(fa_raw, freqs, bw, normalize=True)
 
 
-@pytest.mark.parametrize("use_cython", use_cython)
-def test_smooth(use_cython):
-    try:
-        calc = pykooh.smooth(freqs, freqs, fa_raw, bw, use_cython=use_cython)
-    except ImportError:
-        pytest.skip("Cython is not installed")
-
+def test_smooth():
+    calc = pykooh.smooth(freqs, freqs, fa_raw, bw)
     np.testing.assert_allclose(calc, fa_sm_obspy, rtol=1e-3)
 
 
 def test_smooth_simpl():
-    calc = pykooh.smooth(freqs, freqs, fa_raw, bw, use_cython=False, simplified=True)
+    calc = pykooh.smooth(freqs, freqs, fa_raw, bw, simplified=True)
     mask = np.isfinite(calc)
     np.testing.assert_allclose(calc[mask], fa_sm_obspy[mask], rtol=0.065)
 
@@ -41,18 +34,14 @@ def test_cached_smoother():
     np.testing.assert_allclose(calc, fa_sm_obspy, rtol=1e-3)
 
 
-@pytest.mark.parametrize("use_cython", use_cython)
 @pytest.mark.parametrize(
     "freqs_sm", [np.linspace(0, 50, num=256), np.geomspace(0.1, 50, num=256)]
 )
-def test_same_smoothing(use_cython, freqs_sm):
+def test_same_smoothing(freqs_sm):
     smoother = pykooh.CachedSmoother(freqs, freqs_sm, bw, normalize=True)
     fa_sm_cache = smoother(fa_raw)
 
-    try:
-        fa_sm_calc = pykooh.smooth(freqs_sm, freqs, fa_raw, bw, use_cython=use_cython)
-    except ImportError:
-        pytest.skip("Cython is not installed")
+    fa_sm_calc = pykooh.smooth(freqs_sm, freqs, fa_raw, bw)
 
     # FIXME: Can we use a tigher tolerance
     np.testing.assert_allclose(fa_sm_cache, fa_sm_calc, rtol=0.025)